// backend/src/subscription/enums/operation-key.enum.ts
export enum OperationKey {
    ORDERS_CREATE = 'orders.create',
    SITES_CAPACITY = 'sites.capacity',
    USERS_CAPACITY = 'users.capacity',
    STORAGE_SLOTS_CAPACITY = 'storage_slots.capacity',
    CLIENTS_CREATE = 'clients.create',
    PRINTS_TICKET = 'prints.ticket',
    STORAGE_UPLOAD = 'storage.upload',
}